Clarified how the logic works around 3, 4, 5 byte messages.
Yes, this adds a goto into the code.  I think it's the cleanest
way to indicate what's happening, it's only to a few lines away.
Previously folks had to understand that the additional checksums
would always fall through because the code was zeroed and the
checsum was XOR based.  That's obtuse and confusing.